refactor: Seperate fetching functions
